Modularize interest-island and raise search limit
Extract interest island logic into three services: profileSelector.service.js (profile loading, caching, diversity & matching), suppressionScorer.service.js (suppression signal loading, caching & penalty scoring), and rankingOrchestrator.service.js (scoring, density constraint and ranking pipeline). Trimmed server/util/interestIsland.service.js to delegate to these new modules and kept promotion/upsert hooks. Also increased article search default limit in articleSearch.service.js from 300 to 1000. The refactor separates concerns for readability, testability and reuse while preserving existing ranking behavior.
